Uploading final version of the graphic showing election results in Pennsylvania.
Note that the graphic is not being updated live. The results being used are unofficial vote tallies from the Pennsylvania Department of State and include mail-in and in-person votes in the presidential election. Write-in votes are not included in the DOS count and are therefore not represented in the graphic. Upload includes Python used to scrape the DOS website, resulting CSV file, JSON being used to map Pennsylvanian counties and actual code for the D3 cloropleth map (HTML).